Add prepare for archive build script
As preparations for eZ Platform and part of move to pure composer
based distribution model, we are simplifying of our build environment
and moving the simplified parts here to open them to contributions.

Note: Scripts in a ```.``` folder, like the once in ```.travis``` are considered
"internal" as in they can change at anytime in regards to which
parameters they expect and so on, but this one is fairly small so
will most likely stay similar over time.