[bdd] Disabled demo/clean
The 'clean' install used by this profile isn't compatible with the DemoBundle.

The login screen will use the demobundle's pagelayout, that expects data
from the demobundle. The menu will lead to an exception.

It doesn't show up on Travis since debugging is turned off. But it will fail
in all other environments, or if USE_DEBUGGING is set to 1.