[Docker] Don't specify hard coded default of SYMFONY_DEBUG
Instead of setting default, only pass it if it is defined in env.
This makes sure you can just set SYMFONY_ENV=dev and you are set
with debug, as Symfony will automatically set debug to 1 if in dev.

You can still override SYMFONY_DEBUG if you want, but it's optional.