Let .gitignore behave like a whitelist for /bin folder
Sometimes third party composer packages will bring up new binaries
that are stored beneath `/bin`. By ignoring all files - except the files that
should be part of the project - newly added files are automatically ignored.

To add a file to the project that file needs to be added to the list. This
way consumes less time in maintaining.