EZP-31982: Revert cache setting in Encore config to default
In practice disabled Encore built-in cache on dev environments, following Symfony's default configuration in this matter.
Fixed an issue where adding new entrypoints to webpack.config.js caused them to not be recognized until cache was
cleared - which lead to confusing errors on occuring on templates files that included those entrypoints.